Fixed Bugs
Sparse and DiagonalSparse didnt conform to protocols required for Double and Int multiplication.

TODO:

Need to change TimeIndependantSchrodingerSystem so that it can be defined using any OperatorType,
rather than a Matrix, then changing after. Because for systems like Jaynes-Cummings model, it is
actually less efficient to define the hamiltonian as a Matrix then convert to Sparse or DiagonalSparse,
than it is to run the entire simulation.