Reinstated Parallel - Fixed bug
Bug was that the output of matrix times vector was initialised as Vector, hence
having a single pointer to its memory, so when trying to update its elements in parallel, it was showing a data race.

Fixed by initialising only the elements as an array, updating those, then returning the corresponding vector.