Partial Trace Completed
Finished and tested partial trace algorithm. Tested for two and three qubit systems.

Could look at testing for larger systems but this would require a lot of manual work on paper to do the maths.

Made the function generic so it works on any type conforming to OperatorType.

To do this, added methods to the protocol to get and set with subsrcipts and initialise an empty instance in a vector space.